requirements.txt
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0
boto3>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0 